VM/AWS/requirements.txt
absl-py==0.9.0
ansible==2.9.6
asn1crypto==1.2.0
astor==0.8.1
astroid==2.3.3
beautifulsoup4==4.8.2
boto3==1.12.31
botocore==1.15.31
bs4==0.0.1
cachetools==4.0.0
camera==1.3.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
click==7.1.1
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
dialogflow==0.8.0
dj-database-url==0.5.0
Django==2.0.7
django-heroku==0.3.1
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
gcloud==0.17.0
gensim==3.8.1
geographiclib==1.50
geopy==1.21.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.17.0
idna==2.8
IP2Location==8.4.1
ipaddress==1.0.23
ipdata==3.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
newsapi-python==0.2.6
numpy==1.18.1
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.0
pandas==1.0.3
pexpect==4.8.0
Pillow==7.1.1
praw==6.5.1
prawcore==1.0.1
protobuf==3.11.3
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pycryptodome==3.4.3
pylint==2.4.4
pyOpenSSL==19.0.0
pyparsing==2.4.6
Pyrebase==3.0.27
PySocks==1.7.1
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
smart-open==1.10.0
soupsieve==2.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.36.1
typed-ast==1.4.0
update-checker==0.16
uritemplate==3.0.1
urllib3==1.24.2
websocket-client==0.57.0
Werkzeug==1.0.0
wget==3.2
whitenoise==5.0.1
wrapt==1.11.2
